Restructure internal package names
