--- types/types.go
@@ -1,8 +1,0 @@
-package types
-
-type Package struct {
-	GoPackagePath string
-	ReplacedPath  string
-	Version       string
-	Hash          string
-}
